namespaces are going to be the end of me
