+++ nestjs/src/match/match.controller.ts
@@ -22,14 +22,14 @@ export class MatchController {
 
 	//in the frontend start listening for events on this matchid
 	//then as soon as the match finds an opponent, send an event
-	@Post('find')
-	@UseGuards(No2FAGuard)
-	@UseFilters(UnauthorizedFilter)
-	findMatch(@Req() request, settings: MatchSettings) {
-		Logger.log(`API/MATCH/FIND called`);
-		const id = this.matchService.findMatch(request.user.intra_name, settings);
-		return id;
-	}
+	// @Post('find')
+	// @UseGuards(No2FAGuard)
+	// @UseFilters(UnauthorizedFilter)
+	// findMatch(@Req() request, settings: MatchSettings) {
+	// 	Logger.log(`API/MATCH/FIND called`);
+	// 	const id = this.matchService.findMatch(request.user.intra_name, settings);
+	// 	return id;
+	// }
 
 	@Get('accept/:id')
 	@UseGuards(No2FAGuard)
