+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "bcryptjs": "^2.4.3"
+  }
+}
